Fully support all versions of Armeria
