--- apm-sniffer/apm-sdk-plugin/armeria-0.8.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/armeria/ArmeriaServerInterceptor.java
@@ -1,89 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.skywalking.apm.plugin.armeria;
-
-import com.linecorp.armeria.common.DefaultHttpRequest;
-import com.linecorp.armeria.common.HttpHeaders;
-import io.netty.util.AsciiString;
-import org.apache.skywalking.apm.agent.core.context.CarrierItem;
-import org.apache.skywalking.apm.agent.core.context.ContextCarrier;
-import org.apache.skywalking.apm.agent.core.context.ContextManager;
-import org.apache.skywalking.apm.agent.core.context.tag.Tags;
-import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
-import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
-
-import java.lang.reflect.Method;
-
-/**
- * @author kezhenxu94
- */
-public class ArmeriaServerInterceptor implements InstanceMethodsAroundInterceptor {
-    @Override
-    public void beforeMethod(
-        final EnhancedInstance objInst,
-        final Method method,
-        final Object[] allArguments,
-        final Class<?>[] argumentsTypes,
-        final MethodInterceptResult result) throws Throwable {
-
-        DefaultHttpRequest httpRequest = (DefaultHttpRequest) allArguments[1];
-        HttpHeaders headers = httpRequest.headers();
-
-        ContextCarrier carrier = new ContextCarrier();
-        for (CarrierItem item = carrier.items(); item.hasNext(); ) {
-            item = item.next();
-            item.setHeadValue(headers.get(AsciiString.of(item.getHeadKey())));
-        }
-
-        AbstractSpan entrySpan = ContextManager.createEntrySpan(httpRequest.path(), carrier);
-        entrySpan.setComponent(ComponentsDefine.ARMERIA);
-        entrySpan.setLayer(SpanLayer.HTTP);
-        entrySpan.setPeer(httpRequest.authority());
-        Tags.URL.set(entrySpan, httpRequest.path());
-        Tags.HTTP.METHOD.set(entrySpan, httpRequest.method().name());
-    }
-
-    @Override
-    public Object afterMethod(
-        final EnhancedInstance objInst,
-        final Method method,
-        final Object[] allArguments,
-        final Class<?>[] argumentsTypes,
-        final Object ret) {
-        if (ContextManager.isActive()) {
-            ContextManager.stopSpan();
-        }
-        return ret;
-    }
-
-    @Override
-    public void handleMethodException(
-        final EnhancedInstance objInst,
-        final Method method,
-        final Object[] allArguments,
-        final Class<?>[] argumentsTypes,
-        final Throwable t) {
-        if (ContextManager.isActive()) {
-            ContextManager.activeSpan().errorOccurred().log(t);
-        }
-    }
-}
