+++ apm-sniffer/apm-sdk-plugin/pom.xml
@@ -80,7 +80,7 @@
         <module>ehcache-2.x-plugin</module>
         <module>pulsar-plugin</module>
         <module>netty-socketio-plugin</module>
-        <module>armeria-0.8.x-plugin</module>
+        <module>armeria-0.84.x-plugin</module>
     </modules>
     <packaging>pom</packaging>
