+++ apm-sniffer/optional-plugins/pom.xml
@@ -48,6 +48,7 @@
         <module>zookeeper-3.4.x-plugin</module>
         <module>customize-enhance-plugin</module>
         <module>play-2.x-plugin</module>
+        <module>armeria-0.85.x-plugin</module>
     </modules>
 
     <dependencies>
