+++ test/plugin/scenarios/armeria-0.96minus-scenario/config/expectedData.yaml
@@ -1,0 +1,80 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+registryItems:
+  applications:
+    - {'armeria-0.96minus-scenario': 2}
+  instances:
+    - {'armeria-0.96minus-scenario': 1}
+  operationNames:
+    - 'armeria-0.96minus-scenario': [/greet/skywalking]
+  heartbeat: []
+segmentItems:
+  - applicationCode: armeria-0.96minus-scenario
+    segmentSize: ge 2
+    segments:
+      - segmentId: not null
+        spans:
+          - operationName: /greet/skywalking
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 79
+            componentName: ''
+            isError: false
+            spanType: Entry
+            peer: 127.0.0.1:8085
+            peerId: 0
+            tags:
+              - {key: url, value: /greet/skywalking}
+              - {key: http.method, value: GET}
+      - segmentId: not null
+        spans:
+          - operationName: /greet/skywalking
+            operationId: 0
+            parentSpanId: 0
+            spanId: 1
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 79
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: 127.0.0.1:8085
+            peerId: 0
+            tags:
+              - {key: http.method, value: GET}
+          - operationName: /greet/skywalking
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 1
+            componentName: ''
+            isError: false
+            spanType: Entry
+            peer: ''
+            peerId: 0
+            tags:
+              - {key: url, value: 'http://localhost:8080/greet/skywalking'}
+              - {key: http.method, value: GET}
