+++ test/plugin/scenarios/armeria-0.96minus-scenario/src/main/resources/application.yaml
