--- test/plugin/scenarios/armeria-scenario/config/expectedData.yaml
@@ -1,85 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-registryItems:
-  applications:
-    - {armeria-scenario: 2}
-  instances:
-    - {armeria-scenario: 1}
-  operationNames:
-    - armeria-scenario: [/greet/skywalking]
-  heartbeat: []
-segmentItems:
-  - applicationCode: armeria-scenario
-    segmentSize: ge 4
-    segments:
-      - segmentId: not null
-        spans:
-          - operationName: /greet/skywalking
-            operationId: 0
-            parentSpanId: -1
-            spanId: 0
-            spanLayer: Http
-            startTime: nq 0
-            endTime: nq 0
-            componentId: 79
-            componentName: ''
-            isError: false
-            spanType: Entry
-            peer: 127.0.0.1:8085
-            peerId: 0
-            tags:
-              - {key: url, value: /greet/skywalking}
-              - {key: http.method, value: GET}
-            refs:
-              - {parentEndpointId: 0, parentEndpoint: /greet/skywalking, networkAddressId: 0,
-                 entryEndpointId: 0, refType: CrossProcess, parentSpanId: 1, parentTraceSegmentId: not null,
-                 parentServiceInstanceId: 1, networkAddress: '127.0.0.1:8085', entryEndpoint: /greet/skywalking,
-                 entryServiceInstanceId: 1}
-      - segmentId: not null
-        spans:
-          - operationName: /greet/skywalking
-            operationId: 0
-            parentSpanId: 0
-            spanId: 1
-            spanLayer: Http
-            startTime: nq 0
-            endTime: nq 0
-            componentId: 79
-            componentName: ''
-            isError: false
-            spanType: Exit
-            peer: 127.0.0.1:8085
-            peerId: 0
-            tags:
-              - {key: http.method, value: GET}
-          - operationName: /greet/skywalking
-            operationId: 0
-            parentSpanId: -1
-            spanId: 0
-            spanLayer: Http
-            startTime: nq 0
-            endTime: nq 0
-            componentId: 1
-            componentName: ''
-            isError: false
-            spanType: Entry
-            peer: ''
-            peerId: 0
-            tags:
-              - {key: url, value: 'http://localhost:8080/greet/skywalking'}
-              - {key: http.method, value: GET}
